# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "User Controller",
        "version": "0.1"
    },
    "paths": {
        "/api/test": {
            "get": {
                "tags": [
                    "test"
                ],
                "summary": "get all users",
                "operationId": "index",
                "responses": {
                    "200": {
                        "description": "get all users."
                    }
                }
            },
            "post": {
                "tags": [
                    "test"
                ],
                "summary": "create user",
                "operationId": "store",
                "parameters": [
                    {
                        "name": "name",
                        "in": "query"
                    },
                    {
                        "name": "email",
                        "in": "query"
                    },
                    {
                        "name": "password",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation"
                    },
                    "403": {
                        "description": "forbidden operation"
                    },
                    "422": {
                        "description": "unprocessable entity"
                    }
                }
            }
        },
        "/api/test/{id}": {
            "get": {
                "tags": [
                    "test"
                ],
                "summary": "show user",
                "operationId": "show",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation"
                    },
                    "401": {
                        "description": "unautharize"
                    },
                    "403": {
                        "description": "forbidden operation"
                    },
                    "404": {
                        "description": "user not found"
                    },
                    "422": {
                        "description": "unprocessable entity"
                    }
                }
            },
            "put": {
                "tags": [
                    "test"
                ],
                "summary": "edit user data",
                "operationId": "update",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "name": "name",
                        "in": "query"
                    },
                    {
                        "name": "email",
                        "in": "query"
                    },
                    {
                        "name": "password",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation"
                    },
                    "401": {
                        "description": "unautharize"
                    },
                    "403": {
                        "description": "forbidden operation"
                    },
                    "404": {
                        "description": "user not found"
                    },
                    "422": {
                        "description": "unprocessable entity"
                    }
                }
            },
            "delete": {
                "tags": [
                    "test"
                ],
                "summary": "remove user",
                "operationId": "destroy",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "name": "password",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation"
                    },
                    "401": {
                        "description": "unautharize"
                    },
                    "403": {
                        "description": "forbidden operation"
                    },
                    "404": {
                        "description": "user not found"
                    },
                    "422": {
                        "description": "unprocessable entity"
                    }
                }
            }
        }
    }
}